Refactor project structure into packages
