+++ app/src/main/java/com/example/littlelemon/ui/navigation/Destinations.kt
@@ -1,4 +1,4 @@
-package com.example.littlelemon
+package com.example.littlelemon.ui.navigation
 
 interface Destinations {
     val route : String
