+++ app/src/main/java/com/example/littlelemon/ui/screens/OrderSummary.kt
@@ -1,4 +1,4 @@
-package com.example.littlelemon
+package com.example.littlelemon.ui.screens
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -44,6 +44,8 @@ import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.bumptech.glide.integration.compose.GlideImage
+import com.example.littlelemon.repository.CartRepository
+import com.example.littlelemon.repository.formatDigits
 
 @OptIn(ExperimentalGlideComposeApi::class)
 @Composable
