+++ app/src/main/java/com/example/littlelemon/util/Constants.kt
@@ -1,4 +1,4 @@
-package com.example.littlelemon
+package com.example.littlelemon.util
 
 const val PREF_FIRST_NAME = "firstName"
 const val PREF_LAST_NAME = "lastName"
